rename thread file and class
